requirements.txt
akivymd==1.2.1
DateTime==4.4
Kivy==2.0.0
kivymd==1.1.1
KivyMD-Components==0.1.6
kivymd-extensions==1.0.0
kivymd-extensions.akivymd==1.2.6
multitasking==0.0.11
Pillow==9.0.0
plyer==2.0.0
pyjnius==1.3.0
pymongo==3.12.1
requests==2.23.0